Added the universal platform
This is WIP. In theory, it should work. Basically, this reduces jar size as the Velocity platform no longer shades adventure, and the Spigot platform no longer shades SnakeYaml because in both cases the lib is already bundled with the platform. Shared remains as the core of MCUtils without shading dependencies, while Universal shades all dependencies and Shared, so basically what Shared used to be.