Use bungee serializer instead of adventure bukkit
This reduces jar size significantly. I'm also planning on minimizing all libraries as MCUtils is designed to be used as an all-in-one library, not as a bundle of libraries, meaning that ideally you shouldn't have the need to work with Adventure nor SnakeYaml directly, but only with MCUtils classes instead. For Strings, an easy to use legacy string to Component method will be provided.