Fix lib relocation when lib is not included
Alright, this basically occurred because the shared project used to relocate almost every library, for example, SnakeYaml was being relocated as YamlFile uses it and it's part of the shared project, but when using MCUtils-Spigot as a plugin, an exception was thrown as the relocated classes weren't present. This is because MCUtils-Spigot does not shade SnakeYaml as it is present on Spigot already. Sadly, now every platform has to relocate its libraries to fix this issue (Except jetbrains annotations, which are global and as such part of the mcutils shadow conventions)