Spigot: Fix adventure dependencies
As it turns out, the bungee serializer requires the json and gson serializers to work, this is what was causing the jar to go up ~100KB in size, not the adventure api itself being shaded. MCUtils-Spigot now works and is intended to be used as a plugin instead of being shaded due to its size, though it can still be shaded into your jar file if you don't mind the extra file size.