DataHandler: Added modification status check
This prevents unnecessarily saving anything that relies on a DataMap. Basically DataHandler implementations must check if the DataMap has actually been modified or not before saving, if not, saving can be skipped, if modified, saving can be done and the map must be marked as unmodified in order to save again later on. FlatStorage and YamlFile (And all Yaml extensions) already do this so most developers won't even have to worry about this unless they are making their own DataHandler implementations.