Added the paper platform (See desc)
Alright so this is basically just the spigot platform without any dependencies shaded into it. This is because Paper already includes everything that the MCUtils spigot platform needs as a dependency. This results in a smaller jar size (746KB vs 202KB currently). Paper only projects or paper based servers that download MCUtils as a plugin may benefit from this.