Rename project to SkyUtils (See desc)
About the decision:

Alright, this is a pretty big change. For a pretty long time now I felt like MCUtils was a very generic name, hard to differentiate from other projects. This new name aims to give the project its own identity, as all my projects will now use the "Sky" prefix, at least the ones I do on Minecraft.

Technical information:

This commit has been mostly done with the refractor tool that IntelliJ provides. There is a chance that references to the old name may still exist on comments. Some classes are still prefixed with MC as I will go one by one deciding if they should be renamed or not. Which is determined by what the class should be related with, if the API or Minecraft itself.