MCConsole -> SkyConsole & more flexibility
This allows to extend existing utils classes such as SpigotUtils and VelocityUtils with a custom SkyConsole class type while still only using one field for it instead of two. The reason why I'm using a method and not constructor parameters is for simplicity for those that don't want custom types.