[build-skip] Added gradle publish
Finally, the https://repo.codersky.net maven repository exists, we can now publish releases and snapshots of SkyUtils to it. Right now, we are going to be using "1.0.0-SNAPSHOT" as the version that will be constantly updated until the first release is done. This is mostly because SkyUtils is changing constantly now and making a new spanshot with every change would just add a lot of unnecessary work