SpigotUtils: Added command getters
I didn't find a way to easily get registered commands on velocity, that's the reason why this is only supported on Spigot (And thus paper). Obviously a way to do this would be to store registered commands, but keeping in mind that this is a feature that will most likely be used by very few plugins, I think the extra waste of memory isn't worth it. Plugins that need this most likely will just use a simple workaround of their own.