Gradle setup to shade JSky into SkyUtils
The Codersky repository is currently down due to some issues with the hosting company. However, JSky will be the core of SkyUtils, so I decided to get some work done at least with gradle. The gradle build logic needs a refresh though, as it kind of "works", but not as I would like. Also publishing has some issues with the sources of dependencies