Implemented JSky on shared module (Read desc)
Alright so the progress to finally shade JSky into SkyUtils begins now. The shared module has been adapted to do so. This required heavy refactoring and breaks all other modules, but it will be worth it in the future. Most utility classes have been removed and replaced by their (almost) exact copies present on JSky. The Replacer class is the exception to this rule. SkyReplacer has been created to support adventure, as JSky's Replacer class doesn't support it