SkyMessage system: Message filters
This is the first step for the new SkyMessage system, which will replace the current pattern matching based message system that SkyUtils offers. This system will use JSky's JTag system and will fix the current Adventure relocation problem with Spigot and Paper, as it will use Adventure internally without exposing it to the API. Sending messages as JSON to players and as plain text to the console.