Added basic SkyMessage tests (Paper)
So... I haven't been able to actually test because auth servers seems to be down (My luck is amazing). But this should work, at least the server starts in 1.21.4. The paper platform is probably the one that will be used for shared testing, as it is the fastest to load and doesn't require a proxy.