New module for testing: test-plugins
Alright, so after a bit of research I found that this way may be the best to run a paper server with a test plugin AND the SkyUtils library (As a plugin). Apparently, run-paper is designed to compile the plugin at src, add it to a paper server and run it, which makes sense, as we are not using anything like JUnit, so the /test folder isn't used. With this new structure, test-plugins:paper is a project on itself, and SkyUtils is just compiled and added to the plugins folder for it to work.