SkyMessage system overhaul
Turns out, coloring components can be hard. SkyMessageBuilder is now builds a String directly, this String can then be parsed by SkyMessage#of(String). Now, the process is something like this:
- Filters are applied early. A map is generated with raw strings for each found filter (This works)
- Colors are applied to the contents of each message part (This partially works)
- Events are applied, this creates Components. The message is now considered to be "compiled"

The best part of this system? It's very easy to add new tags and color patterns, and it should be quite fast, as it works with simple Strings up to the last step. Then, a SkyMessage can be reused so no parsing is required. SkyReplacer will be adapted in the future so it can replace at SkyMessages. This leaves us with the possibility to create some sort of "CompiledMessagesFile", where all messages are compiled to SkyMessage and then cached, this would make sending messages from a file incredibly fast with almost zero impact on server resources, as the only things being done at that point are:
- Creating a new Component made out of filtered parts of the message for the receiver.
- If the receiver is a player: Convert to JSON and send.
- If the receiver is the console: Convert to a legacy component and send.