Simplify message filtering. A lot
Ok, so after many, many hours spent trying to make a message system that supports custom filters, I've decided that the cons outweigh the benefits. Here are the cons:
- Extra processing was required just to separate between different "parts"
- A SkyMessage couldn't be fully compiled, as having custom filters leads to infinite ways that the same message can be sent.
- Due to this "infinite" nature, colors needed to be applied after filtering.
- As colors were applied last, events needed to be applied at sending time too.

Pros:
- Custom filters for VERY specific use cases, 99% of users and developers would just use the default player and console filter, and that's it.

In conclusion: The only benefit was a very niche case that can be implemented by a developer in no time. I was thinking on a "hasPerm" filter, but most plugins do a permission check and then send the corresponding message if the user has or hasn't permission, so it was pointless to let users control that, plus you can perfectly use PlaceholderAPI for a permission-based fragment of the message, or even the built-in Replacer class! Solutions already exist, this was a waste of resources and time, but hey, it was fun.